♻️🏢 Prepare dashboard for more
