+++ apps/admin/src/routes/_layout.dashboard/route.tsx
@@ -1,12 +1,10 @@
 import { PageLayout } from "#core/layout/page";
 import { getPageTitle } from "#core/page-title";
 import type { MetaFunction } from "@remix-run/node";
-import { useLoaderData } from "@remix-run/react";
 import { ActiveAnimalsCard } from "./active-animals-card";
 import { AnimalsToSterilizeCard } from "./animals-to-sterilize-card";
 import { AnimalsToVaccinateCard } from "./animals-to-vaccinate-card";
 import { DogsToDiagnoseCard } from "./dogs-to-diagnose-card";
-import type { loader } from "./loader.server";
 import { ManagedAnimalsCard } from "./managed-animals-card";
 
 export { loader } from "./loader.server";
@@ -16,8 +14,6 @@ export const meta: MetaFunction = () => {
 };
 
 export default function Route() {
-  const { isCurrentUserManager } = useLoaderData<typeof loader>();
-
   return (
     <PageLayout.Root>
       <PageLayout.Content className="flex flex-col gap-1 md:gap-2">
@@ -27,7 +23,8 @@ export default function Route() {
           <DogsToDiagnoseCard />
         </section>
 
-        {isCurrentUserManager ? <ManagedAnimalsCard /> : null}
+        <ManagedAnimalsCard />
+
         <ActiveAnimalsCard />
       </PageLayout.Content>
     </PageLayout.Root>
